backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1733215607532/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1666788492850/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1749972191589/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560567968/work
charset-normalizer==3.4.2
cmdstanpy @ file:///home/conda/feedstock_root/build_artifacts/cmdstanpy_1734212843041/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1729602393386/work
convertdate @ file:///home/conda/feedstock_root/build_artifacts/convertdate_1734513016989/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
Cython @ file:///Users/runner/miniforge3/conda-bld/cython_1727455891221/work
deprecation==2.1.0
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1749759981917/work
django-cors-headers==4.7.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1746215177513/work
exceptiongroup==1.3.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1751573414050/work
gotrue==2.12.3
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
holidays @ file:///home/conda/feedstock_root/build_artifacts/holidays_1750157766759/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna==3.10
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1748019130050/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1725459195356/work
matplotlib==3.9.2
munkres==1.1.4
narwhals==1.45.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1668919173192/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1734769783178/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1726878426767/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1630696651849/work
plotly==6.2.0
pmdarima @ file:///Users/runner/miniforge3/conda-bld/pmdarima_1720972235438/work
postgrest==1.1.1
prophet @ file:///Users/runner/miniforge3/conda-bld/prophet_1728043722393/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMeeus @ file:///home/conda/feedstock_root/build_artifacts/pymeeus_1734512290916/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
realtime==2.5.3
requests==2.32.4
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1726082662682/work/dist/scikit_learn-1.5.2-cp39-cp39-macosx_10_13_x86_64.whl#sha256=8b240dbd0bcb2ad759e13f02fa948ad006dc48cdb26d95589456911be0903772
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1716470285601/work/dist/scipy-1.13.1-cp39-cp39-macosx_10_13_x86_64.whl#sha256=c9dd6de2b24ca5604e6e5444642497657db2f4e882c088c0d710addb7781a300
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1734007102193/work
stanio @ file:///home/conda/feedstock_root/build_artifacts/stanio_1734581598953/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1751362449266/work
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1736692542187/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
websockets==15.0.1
xgboost @ file:///Users/runner/miniforge3/conda-bld/xgboost-split_1730232572796/work/python-package
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
gunicorn

# Any other libraries you're using
